backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "tensorflow>=2.19.0",
    "uvicorn>=0.34.3",
]
